Day 13/ShoppingSolution/ShoppingDALLibrary/ProductRepository.cs
using ExceptionHandling;
using ShoppingModalLibrary.cs;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace ShoppingDALLibrary
{
    public class ProductRepository : AbstractRepository<int, Product>
    {
        public override async Task<Product> Delete(int key)
        {
            Product product = await GetByKey(key);
            if (product != null) { items.Remove(product); };
            return product;
            throw new ItemNotFoundException();
        }

        public override async Task<Product> GetByKey(int key)
        {
            for (int i = 0; i < items.Count; i++)
            {
                if (items[i].Id == key)
                    return items[i];
            }
            throw new ItemNotFoundException();
        }

        public override async Task<Product> Update(Product item)
        {
            Product product = await GetByKey(item.Id);
            if (product != null)
            {
                product = item;
                return product;
            }
            throw new ItemNotFoundException();
        }
    }
}
